LineChartComponent is created in order to refactor metrics
